Create LoadStatus abstraction
